+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - Support for Oracle distribution of Hadoop with Big Data service
+- Support for APM Synthetics service added
 
 ## 4.34.0 (July 07, 2021)